// CTT_Online_Shopping/FrontEnd/admin-portal/src/app/models/enums.ts
export enum FilterAttributeType {
  String = 'String',
  Boolean = 'Boolean',
  Number = 'Number',
}

export enum Gender {
  Male = 'Men',
  Female = 'Women',
}
export enum MenCategories {
  Clothing = 'Clothing',
  Accessories = 'Accessories',
}

export enum MenSubategoryClothings {
  TShirts = 'T-Shirts',
  CasualShirts = 'Casual Shirts',
  FormalShirts = 'Formal Shirts',
  Suits = 'Suits',
  Jeans = 'Jeans',
  CasualTrousers = 'Casual Trousers',
  FormalTrousers = 'Formal Trousers',
  Shorts = 'Shorts',
  TrackPants = 'Track Pants',
  SweatersAndSweatshirts = 'Sweaters & Sweatshirts',
  Jackets = 'Jackets',
  BlazersAndCoats = 'Blazers & Coats',
  SportsAndActiveWear = 'Sports & Active Wear',
  IndianAndFestiveWear = 'Indian & Festive Wear',
  InnerwearAndSleepwear = 'Innerwear & Sleepwear',
}
export enum MenSubcategoryAccesstories {
  WatchesAndWearables = 'Watches & Wearables',
  SunglassesAndFrames = 'Sunglasses & Frames',
  BagsAndBackpacks = 'Bags & Backpacks',
  LuggageAndTrolleys = 'Luggage & Trolleys',
  PersonalCareAndGrooming = 'Personal Care & Grooming',
  WalletsAndBelts = 'Wallets & Belts',
  FashionAccessories = 'Fashion Accessories',
}

export enum WomenCategories {
  IndianWesternWear = 'Indian & Western Wear',
  WesternWear = 'WesternWear',
  Accessories = 'Accessories',
}

export enum WomenSubcategoryIndianAndWesternWear {
  KurtasAndSuits = 'Kurtas & Suits',
  KurtisAndTunics = 'Kurtis & Tunics',
  LeggingsSalwarsChuridars = 'Leggings, Salwars & Churidars',
  SkirtsPalazzos = 'Skirts & Palazzos',
  SareesBlouses = 'Sarees & Blouses',
  DressMaterial = 'Dress Material',
  LehengaCholi = 'Lehenga Choli',
  DupattasShawls = 'Dupattas & Shawls',
}
export enum WomenSubcategoryWesternWear {
  DressesJumpsuits = 'Dresses & Jumpsuits',
  TopsTShirtsShirts = 'Tops, T-Shirts & Shirts',
  JeansJeggings = 'Jeans & Jeggings',
  TrousersCapris = 'Trousers & Capris',
  ShortsSkirts = 'Shorts & Skirts',
  Shrugs = 'Shrugs',
  SweatersSweatshirts = 'Sweaters & Sweatshirts',
  JacketsWaistcoats = 'Jackets & Waistcoats',
  CoatsBlazers = 'Coats & Blazers',
}
export enum WomenSubcategoryAccessories {
  WomenWatches = 'Women Watches',
  Analog = 'Analog',
  Chronograph = 'Chronograph',
  Digital = 'Digital',
  AnalogDigital = 'Analog & Digital',
  Sunglasses = 'Sunglasses',
  EyeGlasses = 'Eye Glasses',
  Belt = 'Belt',
}

export enum VariantType {
  Color = 'Color',
  Size = 'Size',
  ColorAndSize = 'ColorAndSize',
}
// enums
export enum DiscountType {
  None = 'None',
  Fixed = 'Fixed',
  Percentage = 'Percentage',
}